requirements.txt
numpy
datasets
pandas
transformers
scikit-learn
torch
python-dotenv